pythonproj/PycharmProjects/SignalAggregator/requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.45.0
anyio==4.8.0
attrs==25.1.0
audioop-lts==0.2.1
babel==2.16.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.9.0
boto3==1.36.6
botocore==1.36.6
browser-use==0.1.19
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
configparser==7.2.0
courlan==1.3.2
dateparser==1.2.0
defusedxml==0.7.1
distro==1.9.0
fastapi==0.115.7
ffmpy==0.5.0
filelock==3.17.0
filetype==1.2.0
fireworks-ai==0.15.12
Flask==3.1.0
frozenlist==1.5.0
fsspec==2024.12.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
gradio==5.9.1
gradio_client==1.5.2
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
html2text==2024.2.26
htmldate==1.9.3
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
httpx-ws==0.7.1
huggingface-hub==0.27.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jusText==3.0.1
langchain==0.3.15
langchain-anthropic==0.3.4
langchain-aws==0.2.11
langchain-core==0.3.37
langchain-fireworks==0.2.6
langchain-google-genai==2.0.8
langchain-mistralai==0.2.7
langchain-ollama==0.2.2
langchain-openai==0.2.14
langchain-text-splitters==0.3.5
langsmith==0.3.1
lxml==5.3.0
lxml_html_clean==0.4.1
MainContentExtractor==0.0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
monotonic==1.6
multidict==6.1.0
numpy==2.2.2
ollama==0.4.7
openai==1.60.1
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pillow==11.1.0
playwright==1.49.1
posthog==3.10.0
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.3
pyaes==1.6.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
pyee==12.0.0
Pygments==2.19.1
pyparsing==3.2.1
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
ruff==0.9.3
s3transfer==0.11.2
safehttpx==0.1.6
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
starlette==0.45.3
Telethon==1.39.0
tenacity==9.0.0
tiktoken==0.8.0
tld==0.13
tokenizers==0.21.0
tomlkit==0.13.2
tqdm==4.67.1
trafilatura==2.0.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
websockets==14.2
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.18.3
zstandard==0.23.0
